refactor: グラフ統計API(/oc/{id}/stats)を専用のApiコントローラへ移設
ルームページ(/oc)の Page コントローラに API メソッド stats() が同居していたが、
これは React グラフが非同期 fetch する JSON API であり、ページ描画とは責務が別。
app/Controllers/Api/ に専用コントローラを置く流儀に合わせ、
OpenChatStatsApiController へ切り出してルートを張り替える（ロジック・出力は不変）。

検証: /oc/{id}/stats の JSON を移設前後で3室バイト一致を確認。
