+++ app/Controllers/Pages/OpenChatPageController.php
@@ -14,11 +14,9 @@ use App\Services\Recommend\RecommendGenarator;
 use App\Services\Recommend\SimilarSizeRoomService;
 use App\Services\StaticData\Dto\StaticTopPageDto;
 use App\Services\StaticData\StaticDataFile;
-use App\Services\Statistics\StatisticsChartArrayService;
 use App\Views\Meta\OcPageMeta;
 use App\Views\Schema\OcPageSchema;
 use App\Views\Schema\PageBreadcrumbsListSchema;
-use App\Services\Statistics\Dto\StatisticsChartDto;
 use App\Views\Classes\CollapseKeywordEnumerationsInterface;
 use App\Views\Classes\Dto\RankingPositionChartArgDtoFactoryInterface;
 use App\Services\Storage\FileStorageInterface;
@@ -166,29 +164,6 @@ class OpenChatPageController
         ));
     }
 
-    /**
-     * 統計チャートデータ（メンバー数推移・期間タブ可用性）を返す。
-     * graph(React) が初回ロード時にこれを fetch する。これにより bot が叩く /oc 本体の
-     * サーバーレンダリングから統計の重い SQLite 読み取り（member/OHLC/順位）を外す。
-     */
-    function stats(
-        StatisticsChartArrayService $statisticsChartArrayService,
-        int $open_chat_id,
-        int $category,
-    ) {
-        $dto = $statisticsChartArrayService->buildStatisticsChartArray(
-            $open_chat_id,
-            $category > 0 ? $category : null
-        );
-        if (!$dto) {
-            $dto = new StatisticsChartDto(
-                (new \DateTime('-1day'))->format('Y-m-d'),
-                (new \DateTime('now'))->format('Y-m-d')
-            );
-        }
-
-        return response($dto);
-    }
 
     private function getAdminDto(int $open_chat_id)
     {
